Moved mesh reader to subfolder
