+++ Common/include/geometry/CPhysicalGeometry.hpp
@@ -28,7 +28,7 @@
 #pragma once
 
 #include "CGeometry.hpp"
-#include "../CMeshReaderFVM.hpp"
+#include "meshreader/CMeshReaderFVM.hpp"
 
 /*!
  * \class CPhysicalGeometry
