+++ Common/src/geometry/meshreader/meson.build
@@ -1,0 +1,5 @@
+common_src += files(['CBoxMeshReaderFVM.cpp',
+                     'CCGNSMeshReaderFVM.cpp',
+                     'CMeshReaderFVM.cpp',
+                     'CRectangularMeshReaderFVM.cpp',
+                     'CSU2ASCIIMeshReaderFVM.cpp'])
